refactor: restructure app/ to canonical layout per ADR-009
Folder layout reset:
- core/settings.py -> core/config.py
- utils/pagination.py -> core/pagination.py; utils/ removed
- middleware/request_logger.py -> core/request_logger.py; middleware/ removed
- enums/ created — one file per spec domain (eav, identity, scheduling,
  clinical, billing, compliance); enum classes extracted from models.py
- models/models.py split into models/<domain>.py per spec domain
- schemas/schemas.py split into schemas/pagination.py and schemas/errors.py
- repositories/ package created (one file per aggregate, populated next commit)
- services/eav_service.py renamed to services/organization_service.py

The Organization model gains @validates() invariants for country, state, and
timezone, plus a from_create() classmethod factory and a deactivate() mutator
(ADR-009 Model-as-entity). AuditLog gains an `outcome` column with a check
constraint; the Alembic migration follows in a later commit.

This commit only restructures and re-imports — the function-based
OrganizationService and call_service_with_audit wrapper still exist; their
class-based replacement and the AuditWriter come in subsequent commits.
