+++ backend/alembic/env.py
@@ -11,10 +11,10 @@ from alembic import context
 from sqlalchemy.ext.asyncio import create_async_engine
 
 from app.core.database import Base
-from app.core.settings import settings
+from app.core.config import settings
 
 # Import all models so they register with Base.metadata
-import app.models.models  # noqa: F401
+import app.models  # noqa: F401
 
 config = context.config
 if config.config_file_name is not None:
