+++ backend/app/enums/__init__.py
@@ -1,0 +1,14 @@
+"""Domain StrEnum classes, organized by spec domain (one file per domain)."""
+
+from app.enums.billing import InsurancePriority as InsurancePriority
+from app.enums.billing import InvoiceStatus as InvoiceStatus
+from app.enums.billing import PayerType as PayerType
+from app.enums.billing import PaymentMethod as PaymentMethod
+from app.enums.billing import PaymentStatus as PaymentStatus
+from app.enums.clinical import NoteFormat as NoteFormat
+from app.enums.clinical import NoteStatus as NoteStatus
+from app.enums.compliance import ConsentStatus as ConsentStatus
+from app.enums.compliance import FormType as FormType
+from app.enums.eav import FieldType as FieldType
+from app.enums.identity import RoleDomain as RoleDomain
+from app.enums.scheduling import SessionStatus as SessionStatus
