+++ backend/app/enums/scheduling.py
@@ -1,0 +1,12 @@
+"""StrEnum classes for the Scheduling domain (SPEC-003)."""
+
+from enum import StrEnum
+
+
+class SessionStatus(StrEnum):
+    SCHEDULED = "SCHEDULED"
+    CONFIRMED = "CONFIRMED"
+    IN_PROGRESS = "IN_PROGRESS"
+    COMPLETED = "COMPLETED"
+    CANCELLED = "CANCELLED"
+    NO_SHOW = "NO_SHOW"
