+++ backend/app/models/__init__.py
@@ -1,0 +1,28 @@
+"""SQLAlchemy ORM models, organized by spec domain (one file per domain)."""
+
+from app.models.billing import ClientInsurance as ClientInsurance
+from app.models.billing import CPTCode as CPTCode
+from app.models.billing import ICDCode as ICDCode
+from app.models.billing import InsurancePayer as InsurancePayer
+from app.models.billing import Invoice as Invoice
+from app.models.billing import InvoiceLineItem as InvoiceLineItem
+from app.models.billing import Payment as Payment
+from app.models.clinical import ClinicalNote as ClinicalNote
+from app.models.compliance import AuditLog as AuditLog
+from app.models.compliance import ClientConsent as ClientConsent
+from app.models.compliance import ConsentType as ConsentType
+from app.models.compliance import Document as Document
+from app.models.compliance import DocumentType as DocumentType
+from app.models.compliance import FormTemplate as FormTemplate
+from app.models.eav import AttributeValue as AttributeValue
+from app.models.eav import EntityAttribute as EntityAttribute
+from app.models.eav import EntityInstance as EntityInstance
+from app.models.eav import EntityType as EntityType
+from app.models.eav import Organization as Organization
+from app.models.identity import Permission as Permission
+from app.models.identity import Person as Person
+from app.models.identity import PersonRole as PersonRole
+from app.models.identity import Role as Role
+from app.models.identity import RolePermission as RolePermission
+from app.models.scheduling import AppointmentType as AppointmentType
+from app.models.scheduling import Session as Session
