+++ backend/app/models/scheduling.py
@@ -1,0 +1,74 @@
+"""ORM models for the Scheduling domain (SPEC-003)."""
+
+import uuid
+from datetime import datetime
+
+from sqlalchemy import (
+    Boolean,
+    DateTime,
+    ForeignKey,
+    Integer,
+    String,
+    Text,
+    text,
+)
+from sqlalchemy import (
+    Enum as SAEnum,
+)
+from sqlalchemy.dialects.postgresql import UUID
+from sqlalchemy.orm import Mapped, mapped_column
+
+from app.core.database import Base, IdMixin, SoftDeleteMixin, TimestampMixin
+from app.enums.scheduling import SessionStatus
+
+
+class AppointmentType(Base, IdMixin, TimestampMixin):
+    __tablename__ = "appointment_types"
+
+    organization_id: Mapped[uuid.UUID] = mapped_column(
+        UUID(as_uuid=True), ForeignKey("organizations.id"), nullable=False
+    )
+    name: Mapped[str] = mapped_column(String, nullable=False)
+    default_duration_minutes: Mapped[int] = mapped_column(Integer, nullable=False)
+    cpt_code_id: Mapped[uuid.UUID | None] = mapped_column(
+        UUID(as_uuid=True), ForeignKey("cpt_codes.id"), nullable=True
+    )
+    is_telehealth: Mapped[bool] = mapped_column(
+        Boolean, nullable=False, server_default=text("false")
+    )
+    is_intake: Mapped[bool] = mapped_column(Boolean, nullable=False, server_default=text("false"))
+    is_active: Mapped[bool] = mapped_column(Boolean, nullable=False, server_default=text("true"))
+
+
+class Session(Base, IdMixin, TimestampMixin, SoftDeleteMixin):
+    __tablename__ = "sessions"
+
+    organization_id: Mapped[uuid.UUID] = mapped_column(
+        UUID(as_uuid=True), ForeignKey("organizations.id"), nullable=False
+    )
+    appointment_type_id: Mapped[uuid.UUID] = mapped_column(
+        UUID(as_uuid=True), ForeignKey("appointment_types.id"), nullable=False
+    )
+    # Bridge rule: must reference an EntityInstance of type provider
+    provider_instance_id: Mapped[uuid.UUID] = mapped_column(
+        UUID(as_uuid=True), ForeignKey("entity_instances.id"), nullable=False
+    )
+    # Bridge rule: must reference an EntityInstance of type client
+    client_instance_id: Mapped[uuid.UUID] = mapped_column(
+        UUID(as_uuid=True), ForeignKey("entity_instances.id"), nullable=False
+    )
+    start_time: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False)
+    end_time: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False)
+    status: Mapped[SessionStatus] = mapped_column(
+        SAEnum(SessionStatus, native_enum=False),
+        nullable=False,
+        server_default=SessionStatus.SCHEDULED,
+    )
+    cancellation_reason: Mapped[str | None] = mapped_column(Text, nullable=True)
+    cancelled_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
+    cancelled_by_person_id: Mapped[uuid.UUID | None] = mapped_column(
+        UUID(as_uuid=True), ForeignKey("people.id"), nullable=True
+    )
+    location: Mapped[str | None] = mapped_column(String, nullable=True)
+    # Internal scheduling notes — not PHI, not a clinical note
+    notes: Mapped[str | None] = mapped_column(Text, nullable=True)
