+++ backend/app/schemas/__init__.py
@@ -1,0 +1,5 @@
+"""Pydantic schemas — boundary contracts for routers (ADR-009).
+
+One file per domain (e.g. ``schemas/eav.py``) plus shared-purpose files
+(``pagination.py``, ``errors.py``).
+"""
