+++ backend/tests/test_cross_cutting/test_pagination.py
@@ -19,9 +19,7 @@ from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.core.exceptions import BadRequestError
-from app.models.models import Organization
-from app.schemas.schemas import PaginationParams, SortDir
-from app.utils.pagination import (
+from app.core.pagination import (
     apply_date_range_filter,
     apply_exact_filter,
     apply_in_filter,
@@ -30,6 +28,8 @@ from app.utils.pagination import (
     encode_cursor,
     paginate,
 )
+from app.models.eav import Organization
+from app.schemas.pagination import PaginationParams, SortDir
 
 # ===========================================================================
 # Part 1 — Unit tests
